Backend/requirements.txt
fastapi==0.109.0
email-validator==2.1.0
uvicorn[standard]==0.27.0
python-socketio==5.10.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
deepface==0.0.89
tf-keras==2.15.0
opencv-python==4.9.0.80
numpy==1.26.3
geopy==2.4.1
python-dotenv==1.0.0
Pillow==10.2.0
aiofiles==23.2.1
